Import refactored WDKClient's IndeterminateCheckbox into DataGrid
