+++ src/components/grids/DataGrid/index.tsx
@@ -7,17 +7,16 @@ import {
   useRowSelect,
   Cell,
   Row,
-  Hooks,
   Column,
 } from 'react-table';
-import { pickBy, ceil, merge } from 'lodash';
+import { ceil, merge } from 'lodash';
 
 // Definitions
 import stylePresets, { DataGridStyleSpec } from './stylePresets';
 
 // Components
 import { H3 } from '../../typography';
-import IndeterminateCheckbox from './IndeterminateCheckbox';
+import IndeterminateCheckbox from '../../inputs/SelectTree/IndeterminateCheckbox';
 import DataCell from './DataCell';
 import HeaderCell from './HeaderCell';
 import PaginationControls from './PaginationControls';
@@ -171,28 +170,39 @@ export default function DataGrid({
             id: 'selection',
             // The header can use the table's getToggleAllRowsSelectedProps method
             // to render a checkbox
-            Header: ({ getToggleAllPageRowsSelectedProps }) => (
+            Header: ({ getToggleAllPageRowsSelectedProps }) => {
+              const props = getToggleAllPageRowsSelectedProps();
+              return (
               <div>
                 <IndeterminateCheckbox
-                  {...getToggleAllPageRowsSelectedProps()}
-                  themeRole={themeRole}
+                  checked={props.checked ?? false}
+                  indeterminate={props.indeterminate ?? false}
+                  onChange={props.onChange}
+                  value={''}
+                  name={''}
+                  // themeRole={themeRole}
                 />
               </div>
-            ),
+            )},
             // The cell can use the individual row's getToggleRowSelectedProps method
             // to the render a checkbox.
-	    // The `checked` prop returned by getToggleRowSelectedProps is not fit for purpose
-	    // because it only considers initial pre-selected/checked state (originalData[].isSelected).
-	    // Luckily `row.isSelected` has the correct "live" state.
-            Cell: ({ row }) => (
+            // The `checked` prop returned by getToggleRowSelectedProps is not fit for purpose
+            // because it only considers initial pre-selected/checked state (originalData[].isSelected).
+            // Luckily `row.isSelected` has the correct "live" state.
+            Cell: ({ row }) => {
+              const props = row.getToggleRowSelectedProps();
+              return (
               <div>
                 <IndeterminateCheckbox
-                  {...row.getToggleRowSelectedProps()}
-	          checked={row.isSelected}
-                  themeRole={themeRole}
+	                checked={row.isSelected}
+                  indeterminate={props.indeterminate ?? false}
+                  onChange={props.onChange}
+                  name={''}
+                  value={row.id}
+                  // themeRole={themeRole}
                 />
               </div>
-            ),
+            )},
           },
           ...columns,
         ]);
